Rename to optimizer_registry
